backend/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.0
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Auth & Security
pyjwt>=2.8.0
bcrypt>=4.1.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Rate Limiting & Caching
slowapi>=0.1.9
redis>=5.0.0

# Observability
structlog>=23.2.0
python-json-logger>=2.0.7

# Utils
httpx>=0.25.0
python-dotenv>=1.0.0
